Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.

+++ include/fractol.h
@@ -135,7 +135,7 @@ typedef struct s_fractal
 	t_complex			c;
 	double				mu;
 	int					bind_combo;
-	int					modify_color;
+	double					modify_color;
 	int					palette_n;
 	double				power;
 	int					psychedelic_colors;
